add enum to game class. which is now all that's needed
