--- Models/Games/BracketGame.cs
@@ -1,7 +1,0 @@
-namespace CribblyBackend.Models
-{
-    public class BracketGame : Game
-    {
-        public string GameRound { get; set; }
-    }
-}
