--- Models/Games/PlayInGame.cs
@@ -1,7 +1,0 @@
-namespace CribblyBackend.Models
-{
-    public class PlayInGame : Game
-    {
-        public int GameRound { get; set; }
-    }
-}
